+++ resources/views/components/color-picker/index.blade.php
@@ -17,7 +17,7 @@ $placeholderText = $placeholder ?? __('spire::spire-ui.color-picker.placeholder'
     x-data="spireColorPicker({
         @if($wireConfig->hasWireModel())
             value: $wire.entangle('{{ $wireConfig->wireModel }}', {{ $wireConfig->liveModifier() }}),
-        @else
+        @elseif(!$attributes->has('x-model'))
             value: '',
         @endif
         placeholder: '{{ $placeholderText }}',