+++ resources/views/components/select/index.blade.php
@@ -47,7 +47,7 @@ $maxValue = is_numeric($max) ? (int) $max : 'null';
     x-data="spireSelect({
         @if($wireConfig->hasWireModel())
             value: $wire.entangle('{{ $wireConfig->wireModel }}', {{ $wireConfig->liveModifier() }}),
-        @else
+        @elseif(!$attributes->has('x-model'))
             value: {{ $multiple ? '[]' : "''" }},
         @endif
         placeholder: '{{ $placeholderText }}',